Add support for NFS4 ACL
This commit brings optional support to NFS4 ACL. This support is
enabled by setting the following parameters in the [config] section
of the configuration file:

  acl=True (default)
  acl_type=nfs4
  acl_nfs4_domain=example.tld

When enabled, the permissions set on home (and optionally scratch)
directories are 'rwaxDtncy'. It is fully explained in code comments.

The default value of acl_type parameter is posix. When this parameter
is not defined, the script setups POSIX ACL, just like before. Then,
this new feature does not change behaviour on existing installation.