Add possibility to not overwrite user dirs
Add overwrite boolean parameter to control the behaviour of the PAM
module on existing user directories. If True, the module will ensure
home_dir and scratch_dir are directories and have module defined
permissions. If False, the module will not modify home_dir and
scratch_dir if they exist.

The behaviour of the module is not altered by this parameter on
inexisting {user,scratch}_dir.

Default value is True, it corresponds to the previous behaviour.